rebase + cleanup
